requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Upbit API
PyJWT==2.8.0
requests==2.31.0

# Data processing
pandas==2.1.3
numpy==1.26.4
scikit-learn==1.3.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.12.1

# LLM and Document Processing
openai==1.3.7
langchain==0.0.340
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
aiohttp==3.9.1

# Technical Analysis
TA-Lib==0.6.5
scipy==1.11.4

# Monitoring and Logging
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
python-dotenv==1.0.0
black==23.11.0
isort==5.12.0